Add details toggles for FAQs
As requested in #72 this introduces the custom block `::: details` which can be used to initially hide long-form content and first present a summary. In this case the FAQ leverage it to better structure the questions and answers and provide a better overview than the previously used table of contents.

Thoughts: Even though I really like the output, my concern about this is that we need to use a custom block for that and the authoring of the markdown does not seem intuitive. IMHO the headlines plus plaragraphs feel more natural. Feedback welcome!

If merged, this would close #72.

Signed-off-by: Dennis Reimann <mail@dennisreimann.de>