Add currentVersion variable
This substitutes the `{{ $page.currentVersion }}` with whatever is defined in the configs `extendPageData` function. The goal is to have a single place to maintain the current Wasabi version number as it is used in multiple places. This could also be extended for other cases, but I'm putting it out here to discuss the idea.

Note: I haven't found a way to make the substitution work in fenced code blocks yet. that's why this commit uses the `<pre><code>` fallback for code blocks on the Deterministic Builds page. It's ugly, nevertheless worth it imho.

Feedback appreciated!